requirements.txt
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
pandas>=2.0.0
openpyxl>=3.1.0
rapidfuzz>=3.0.0
watchdog>=3.0.0
flask>=3.0.0
